+++ src/app/(tabs)/_layout.tsx
@@ -1,11 +1,10 @@
 import FontAwesome from '@expo/vector-icons/FontAwesome';
-import { Link, Tabs } from 'expo-router';
+import { Tabs } from 'expo-router';
 import React from 'react';
-import { Pressable } from 'react-native';
 
 import Colors from '../../../constants/Colors';
-import { useClientOnlyValue } from '../../components/use-client-only-value';
-import { useColorScheme } from '../../components/useColorScheme';
+import { useClientOnlyValue } from '../use-client-only-value';
+import { useColorScheme } from '../useColorScheme';
 
 // You can explore the built-in icon families and icons on the web at https://icons.expo.fyi/
 function TabBarIcon(props: {
@@ -30,29 +29,18 @@ export default function TabLayout() {
       <Tabs.Screen
         name='index'
         options={{
-          title: 'Tab One',
-          tabBarIcon: ({ color }) => <TabBarIcon name='code' color={color} />,
-          headerRight: () => (
-            <Link href='/modal' asChild>
-              <Pressable>
-                {({ pressed }) => (
-                  <FontAwesome
-                    name='info-circle'
-                    size={25}
-                    color={Colors[colorScheme ?? 'light'].text}
-                    style={{ marginRight: 15, opacity: pressed ? 0.5 : 1 }}
-                  />
-                )}
-              </Pressable>
-            </Link>
-          ),
+          title: 'Map View',
+          headerTitle: 'Happy Hour',
+          tabBarIcon: ({ color }) => <TabBarIcon name='map' color={color} />,
         }}
       />
       <Tabs.Screen
         name='two'
         options={{
-          title: 'Tab Two',
-          tabBarIcon: ({ color }) => <TabBarIcon name='code' color={color} />,
+          title: 'List View',
+          tabBarIcon: ({ color }) => (
+            <TabBarIcon name='list-ul' color={color} />
+          ),
         }}
       />
     </Tabs>
