+++ src/app/(tabs)/two.tsx
@@ -1,35 +1,5 @@
-import { StyleSheet } from 'react-native';
-
-import MapScreen from '../../components/map-screen';
-import { Text, View } from '../../components/themed';
+import { ListScreen } from 'src/features/list';
 
 export default function TabTwoScreen() {
-  return (
-    <View style={styles.container}>
-      <Text style={styles.title}>Tab Two</Text>
-      <View
-        style={styles.separator}
-        lightColor='#eee'
-        darkColor='rgba(255,255,255,0.1)'
-      />
-      <MapScreen />
-    </View>
-  );
+  return <ListScreen />;
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  title: {
-    fontSize: 20,
-    fontWeight: 'bold',
-  },
-  separator: {
-    marginVertical: 30,
-    height: 1,
-    width: '80%',
-  },
-});
