+++ src/features/list/components/list-screen.tsx
@@ -1,0 +1,26 @@
+import { StyleSheet, Text, View } from 'react-native';
+
+export const ListScreen = () => {
+  return (
+    <View style={styles.container}>
+      <View style={styles.separator} />
+      <Text>Carlos O'Bryan's</Text>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+  },
+  title: {
+    fontSize: 20,
+    fontWeight: 'bold',
+  },
+  separator: {
+    marginVertical: 30,
+    height: 1,
+    width: '80%',
+  },
+});
